requirements.txt
pandas
numpy
scikit-learn
nltk
fastapi
uvicorn
mysql-connector-python
sqlalchemy 
pymysql
python-dotenv
